+++ app/auth.py
@@ -10,12 +10,24 @@ from flask import jsonify, request
 # from flask_login import login_required
 from flask_httpauth import HTTPBasicAuth
 from jwt import InvalidTokenError
+from werkzeug.security import generate_password_hash, check_password_hash
 
 # Local modules
 from app.helpers import fetch_query_params
 
 # Basic Authentication - User and Password
 auth = HTTPBasicAuth()
+
+users = {
+    "john": generate_password_hash("hello"),
+    "susan": generate_password_hash("bye")
+}
+
+
+@auth.verify_password
+def verify_password(username, password):
+    if username in users and check_password_hash(users.get(username), password):
+        return username
 
 
 def token_required(f):
